pyproject.toml
[tool.poetry]
name = "recorder"
version = "0.1.0"
description = "A service to continuously record sensor data from an MQTT broker"
authors = ["Justin Arista Kasin <justin+github@aristakas.in>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^1.5.1"
faunadb = "^4.1.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^21.5b0"
mypy = "^0.812"
pylint = "^2.8.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
